// src/app/modules/business-common/src/lib/components/journey-selection/enums/translation-keys.enum.ts
export enum TranslationKeys {
  City_BOG = 'City.BOG',
  Fare_Chargeable_IconLabel = 'Fare.Chargeable_IconLabel',
  Fare_Included_IconLabel = 'Fare.Included_IconLabel',
  Fare_NotIncluded_IconLabel = 'Fare.NotIncluded_IconLabel',
  Schedule_Compare_Fares = 'Schedule.Compare_Fares',
  Schedule_Connection_StopLabel = 'Schedule.Connection_StopLabel',
  Schedule_ExtraDay_Days_Label = 'Schedule.ExtraDay.Days_Label',
  Schedule_Fares_Title_Vertical = 'Schedule.Fares.Title_Vertical',
  Schedule_Stop = 'Schedule.Stop',
  Schedule_Stops = 'Schedule.Stops',
}